# Conflicts:
#	nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/ApplicationApprovalService_New.java
#	nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/gobackdirectly/service/GoBackDirectlyRegisterDefault.java

+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.app/src/main/java/nts/uk/ctx/at/request/app/command/application/common/UpdateApplicationCommonCmd.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class UpdateApplicationCommonCmd {
-	private Long version;
+	private int version;
 	
 	/**
 	 * application ID
