+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/common/service/detailscreen/ProcessCancel.java
@@ -5,5 +5,5 @@ package nts.uk.ctx.at.request.dom.application.common.service.detailscreen;
  *
  */
 public interface ProcessCancel {
-	public void detailScreenCancelProcess(String companyID,String appID, Long version);
+	public void detailScreenCancelProcess(String companyID,String appID, int version);
 }
