+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/common/service/detailscreen/after/AfterProcessDelete.java
@@ -5,6 +5,6 @@ import nts.uk.ctx.at.request.dom.application.common.service.detailscreen.output.
 
 public interface AfterProcessDelete {
 
-	public ProcessDeleteResult screenAfterDelete (String companyID,String appID, Long version );
+	public ProcessDeleteResult screenAfterDelete (String companyID,String appID, int version );
 
 }
