+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/common/service/detailscreen/before/BeforeProcessDenialImpl.java
@@ -14,7 +14,7 @@ public class BeforeProcessDenialImpl implements BeforeProcessDenial {
 	private DetailBeforeUpdate detailBeforeProcessRegisterRepo;
 
 	@Override
-	public void detailedScreenProcessBeforeDenial(String companyId, String appId, Long version) {
+	public void detailedScreenProcessBeforeDenial(String companyId, String appId, int version) {
 		detailBeforeProcessRegisterRepo.exclusiveCheck(companyId, appId, version);
 		
 	}
